Module wayland_protocols_plasma::dpms::server::org_kde_kwin_dpms

Dpms for a wl_output

This interface provides information about the VESA DPMS state for a wl_output. It gets created through the request get on the org_kde_kwin_dpms_manager interface.

On creating the resource the server will push whether DPSM is supported for the output, the currently used DPMS state and notifies the client through the done event once all states are pushed. Whenever a state changes the set of changes is committed with the done event.

Structs

OrgKdeKwinDpms

Dpms for a wl_output

Enums

Event

Mode

Request

Constants

EVT_DONE_OPCODE

The wire opcode for this event

EVT_DONE_SINCE

The minimal object version supporting this event

EVT_MODE_OPCODE

The wire opcode for this event

EVT_MODE_SINCE

The minimal object version supporting this event

EVT_SUPPORTED_OPCODE

The wire opcode for this event

EVT_SUPPORTED_SINCE

The minimal object version supporting this event

REQ_RELEASE_OPCODE

The wire opcode for this request

REQ_RELEASE_SINCE

The minimal object version supporting this request

REQ_SET_OPCODE

The wire opcode for this request

REQ_SET_SINCE

The minimal object version supporting this request