wayland_protocols_plasma::external_brightness::v1::client

Module kde_external_brightness_device_v1

brightness control device

After creating this object, the client should issue all relevant setup requests (set_internal, set_edid, set_max_brightness, optionally set_observed_brightness) and finish the sequence with commit. Afterwards, for each change in values, the client must call commit again.

Structs

KdeExternalBrightnessDeviceV1

brightness control device

Enums

Event

Request

Constants

EVT_REQUESTED_BRIGHTNESS_OPCODE

The wire opcode for this event

EVT_REQUESTED_BRIGHTNESS_SINCE

The minimal object version supporting this event

REQ_COMMIT_OPCODE

The wire opcode for this request

REQ_COMMIT_SINCE

The minimal object version supporting this request

REQ_DESTROY_OPCODE

The wire opcode for this request

REQ_DESTROY_SINCE

The minimal object version supporting this request

REQ_SET_EDID_OPCODE

The wire opcode for this request

REQ_SET_EDID_SINCE

The minimal object version supporting this request

REQ_SET_INTERNAL_OPCODE

The wire opcode for this request

REQ_SET_INTERNAL_SINCE

The minimal object version supporting this request

REQ_SET_MAX_BRIGHTNESS_OPCODE

The wire opcode for this request

REQ_SET_MAX_BRIGHTNESS_SINCE

The minimal object version supporting this request

REQ_SET_OBSERVED_BRIGHTNESS_OPCODE

The wire opcode for this request

REQ_SET_OBSERVED_BRIGHTNESS_SINCE

The minimal object version supporting this request