Module wayland_protocols_plasma::fake_input::client::org_kde_kwin_fake_input

Fake input manager

This interface allows other processes to provide fake input events. Purpose is on the one hand side to provide testing facilities like XTest on X11. But also to support use case like kdeconnect’s mouse pad interface.

A compositor should not trust the input received from this interface. Clients should not expect that the compositor honors the requests from this interface.

Structs

OrgKdeKwinFakeInput

Fake input manager

Enums

Event

Request

Constants

REQ_AUTHENTICATE_OPCODE

The wire opcode for this request

REQ_AUTHENTICATE_SINCE

The minimal object version supporting this request

REQ_AXIS_OPCODE

The wire opcode for this request

REQ_AXIS_SINCE

The minimal object version supporting this request

REQ_BUTTON_OPCODE

The wire opcode for this request

REQ_BUTTON_SINCE

The minimal object version supporting this request

REQ_KEYBOARD_KEY_OPCODE

The wire opcode for this request

REQ_KEYBOARD_KEY_SINCE

The minimal object version supporting this request

REQ_POINTER_MOTION_ABSOLUTE_OPCODE

The wire opcode for this request

REQ_POINTER_MOTION_ABSOLUTE_SINCE

The minimal object version supporting this request

REQ_POINTER_MOTION_OPCODE

The wire opcode for this request

REQ_POINTER_MOTION_SINCE

The minimal object version supporting this request

REQ_TOUCH_CANCEL_OPCODE

The wire opcode for this request

REQ_TOUCH_CANCEL_SINCE

The minimal object version supporting this request

REQ_TOUCH_DOWN_OPCODE

The wire opcode for this request

REQ_TOUCH_DOWN_SINCE

The minimal object version supporting this request

REQ_TOUCH_FRAME_OPCODE

The wire opcode for this request

REQ_TOUCH_FRAME_SINCE

The minimal object version supporting this request

REQ_TOUCH_MOTION_OPCODE

The wire opcode for this request

REQ_TOUCH_MOTION_SINCE

The minimal object version supporting this request

REQ_TOUCH_UP_OPCODE

The wire opcode for this request

REQ_TOUCH_UP_SINCE

The minimal object version supporting this request