wayland_protocols_plasma::fullscreen_shell::client
Module _wl_fullscreen_shell

Displays a single surface per output

Displays a single surface per output.

This interface provides a mechanism for a single client to display simple full-screen surfaces. While there technically may be multiple clients bound to this interface, only one of those clients should be shown at a time.

To present a surface, the client uses either the present_surface or present_surface_for_mode requests. Presenting a surface takes effect on the next wl_surface.commit. See the individual requests for details about scaling and mode switches.

The client can have at most one surface per output at any time. Requesting a surface be presented on an output that already has a surface replaces the previously presented surface. Presenting a null surface removes its content and effectively disables the output. Exactly what happens when an output is “disabled” is compositor-specific. The same surface may be presented on multiple outputs simultaneously.

Once a surface is presented on an output, it stays on that output until either the client removes it or the compositor destroys the output. This way, the client can update the output’s contents by simply attaching a new buffer.

Structs

WlFullscreenShell

Displays a single surface per output

Enums

Capability

capabilities advertised by the compositor

Error

wl_fullscreen_shell error values

Event

PresentMethod

different method to set the surface fullscreen

Request

Constants

EVT_CAPABILITY_OPCODE

The wire opcode for this event

EVT_CAPABILITY_SINCE

The minimal object version supporting this event

REQ_PRESENT_SURFACE_FOR_MODE_OPCODE

The wire opcode for this request

REQ_PRESENT_SURFACE_FOR_MODE_SINCE

The minimal object version supporting this request

REQ_PRESENT_SURFACE_OPCODE

The wire opcode for this request

REQ_PRESENT_SURFACE_SINCE

The minimal object version supporting this request

REQ_RELEASE_OPCODE

The wire opcode for this request

REQ_RELEASE_SINCE

The minimal object version supporting this request