Struct wayland_protocols_plasma::output_device::v1::server::org_kde_kwin_outputdevice::OrgKdeKwinOutputdevice

pub struct OrgKdeKwinOutputdevice { /* private fields */ }

output configuration representation

An outputdevice describes a display device available to the compositor. outputdevice is similar to wl_output, but focuses on output configuration management.

A client can query all global outputdevice objects to enlist all available display devices, even those that may currently not be represented by the compositor as a wl_output.

The client sends configuration changes to the server through the outputconfiguration interface, and the server applies the configuration changes to the hardware and signals changes to the outputdevices accordingly.

This object is published as global during start up for every available display devices, or when one later becomes available, for example by being hotplugged via a physical connector.

This interface has no requests.

Implementations

impl OrgKdeKwinOutputdevice

pub fn geometry(
&self,
x: i32,
y: i32,
physical_width: i32,
physical_height: i32,
subpixel: i32,
make: String,
model: String,
transform: i32
)

geometric properties of the output

The geometry event describes geometric properties of the output. The event is sent when binding to the output object and whenever any of the properties change.

pub fn mode(&self, flags: u32, width: i32, height: i32, refresh: i32, mode_id: i32)

advertise available output modes and current one

The mode event describes an available mode for the output.

When the client binds to the outputdevice object, the server sends this event once for every available mode the outputdevice can be operated by.

There will always be at least one event sent out on initial binding, which represents the current mode.

Later on if an output changes its mode the event is sent again, whereby this event represents the mode that has now become current. In other words, the current mode is always represented by the latest event sent with the current flag set.

The size of a mode is given in physical hardware units of the output device. This is not necessarily the same as the output size in the global compositor space. For instance, the output may be scaled, as described in org_kde_kwin_outputdevice.scale, or transformed, as described in org_kde_kwin_outputdevice.transform.

The id can be used to refer to a mode when calling set_mode on an org_kde_kwin_outputconfiguration object.

pub fn done(&self)

sent all information about output

This event is sent after all other properties have been sent on binding to the output object as well as after any other output property change have been applied later on. This allows to see changes to the output properties as atomic, even if multiple events successively announce them.

pub fn scale(&self, factor: i32)

output scaling properties

This event contains scaling geometry information that is not in the geometry event. It may be sent after binding the output object or if the output scale changes later. If it is not sent, the client should assume a scale of 1.

A scale larger than 1 means that the compositor will automatically scale surface buffers by this amount when rendering. This is used for high resolution displays where applications rendering at the native resolution would be too small to be legible.

It is intended that scaling aware clients track the current output of a surface, and if it is on a scaled output it should use wl_surface.set_buffer_scale with the scale of the output. That way the compositor can avoid scaling the surface, and the client can supply a higher detail image.

pub fn edid(&self, raw: String)

advertise EDID data for the output

The edid event encapsulates the EDID data for the outputdevice.

The event is sent when binding to the output object. The EDID data may be empty, in which case this event is sent anyway. If the EDID information is empty, you can fall back to the name et al. properties of the outputdevice.

pub fn enabled(&self, enabled: i32)

output is enabled or disabled

The enabled event notifies whether this output is currently enabled and used for displaying content by the server. The event is sent when binding to the output object and whenever later on an output changes its state by becoming enabled or disabled.

pub fn uuid(&self, uuid: String)

A unique id for this outputdevice

The uuid can be used to identify the output. It’s controlled by the server entirely. The server should make sure the uuid is persistent across restarts. An empty uuid is considered invalid.

pub fn scalef(&self, factor: f64)

output scaling properties

This event contains scaling geometry information that is not in the geometry event. It may be sent after binding the output object or if the output scale changes later. If it is not sent, the client should assume a scale of 1.

A scale larger than 1 means that the compositor will automatically scale surface buffers by this amount when rendering. This is used for high resolution displays where applications rendering at the native resolution would be too small to be legible.

It is intended that scaling aware clients track the current output of a surface, and if it is on a scaled output it should use wl_surface.set_buffer_scale with the scale of the output. That way the compositor can avoid scaling the surface, and the client can supply a higher detail image.

wl_output will keep the output scale as an integer. In every situation except configuring the window manager you want to use that.

pub fn colorcurves(&self, red: Vec<u8>, green: Vec<u8>, blue: Vec<u8>)

output color curves

Describes the color intensity profile of the output. Commonly used for gamma/color correction.

The array contains all color ramp values of the output. For example on 8bit screens there are 256 of them.

The array elements are unsigned 16bit integers.

pub fn serial_number(&self, serialNumber: String)

Serial Number

Serial ID of the monitor, sent on startup before the first done event.

pub fn eisa_id(&self, eisaId: String)

EISA ID

EISA ID of the monitor, sent on startup before the first done event.

pub fn capabilities(&self, flags: Capability)

capability flags

What capabilities this device has, sent on startup before the first done event.

pub fn overscan(&self, overscan: u32)

overscan

Overscan value of the monitor in percent, sent on startup before the first done event.

pub fn vrr_policy(&self, vrr_policy: VrrPolicy)

Variable Refresh Rate Policy

What policy the compositor will employ regarding its use of variable refresh rate.