Module wayland_protocols_plasma::output_device::v2::client::kde_output_device_v2

output configuration representation

An output device describes a display device available to the compositor. output_device is similar to wl_output, but focuses on output configuration management.

A client can query all global output_device objects to enlist all available display devices, even those that may currently not be represented by the compositor as a wl_output.

The client sends configuration changes to the server through the outputconfiguration interface, and the server applies the configuration changes to the hardware and signals changes to the output devices accordingly.

This object is published as global during start up for every available display devices, or when one later becomes available, for example by being hotplugged via a physical connector.

Structs

Capability

describes capabilities of the outputdevice

KdeOutputDeviceV2

output configuration representation

Enums

Event

Request

RgbRange

describes RGB range policy

Subpixel

subpixel geometry information

Transform

transform from framebuffer to output

VrrPolicy

describes vrr policy

Constants

EVT_CAPABILITIES_OPCODE

The wire opcode for this event

EVT_CAPABILITIES_SINCE

The minimal object version supporting this event

EVT_CURRENT_MODE_OPCODE

The wire opcode for this event

EVT_CURRENT_MODE_SINCE

The minimal object version supporting this event

EVT_DONE_OPCODE

The wire opcode for this event

EVT_DONE_SINCE

The minimal object version supporting this event

EVT_EDID_OPCODE

The wire opcode for this event

EVT_EDID_SINCE

The minimal object version supporting this event

EVT_EISA_ID_OPCODE

The wire opcode for this event

EVT_EISA_ID_SINCE

The minimal object version supporting this event

EVT_ENABLED_OPCODE

The wire opcode for this event

EVT_ENABLED_SINCE

The minimal object version supporting this event

EVT_GEOMETRY_OPCODE

The wire opcode for this event

EVT_GEOMETRY_SINCE

The minimal object version supporting this event

EVT_MODE_OPCODE

The wire opcode for this event

EVT_MODE_SINCE

The minimal object version supporting this event

EVT_NAME_OPCODE

The wire opcode for this event

EVT_NAME_SINCE

The minimal object version supporting this event

EVT_OVERSCAN_OPCODE

The wire opcode for this event

EVT_OVERSCAN_SINCE

The minimal object version supporting this event

EVT_RGB_RANGE_OPCODE

The wire opcode for this event

EVT_RGB_RANGE_SINCE

The minimal object version supporting this event

EVT_SCALE_OPCODE

The wire opcode for this event

EVT_SCALE_SINCE

The minimal object version supporting this event

EVT_SERIAL_NUMBER_OPCODE

The wire opcode for this event

EVT_SERIAL_NUMBER_SINCE

The minimal object version supporting this event

EVT_UUID_OPCODE

The wire opcode for this event

EVT_UUID_SINCE

The minimal object version supporting this event

EVT_VRR_POLICY_OPCODE

The wire opcode for this event

EVT_VRR_POLICY_SINCE

The minimal object version supporting this event