wayland_protocols_plasma::output_management::v1::client
Module org_kde_kwin_outputconfiguration

configure single output devices

outputconfiguration is a client-specific resource that can be used to ask the server to apply changes to available output devices.

The client receives a list of output devices from the registry. When it wants to apply new settings, it creates a configuration object from the outputmanagement global, writes changes through this object’s enable, scale, transform and mode calls. It then asks the server to apply these settings in an atomic fashion, for example through Linux’ DRM interface.

The server signals back whether the new settings have applied successfully or failed to apply. outputdevice objects are updated after the changes have been applied to the hardware and before the server side sends the applied event.

Structs

OrgKdeKwinOutputconfiguration

configure single output devices

Enums

Event

Request

VrrPolicy

describes vrr policy

Constants

EVT_APPLIED_OPCODE

The wire opcode for this event

EVT_APPLIED_SINCE

The minimal object version supporting this event

EVT_FAILED_OPCODE

The wire opcode for this event

EVT_FAILED_SINCE

The minimal object version supporting this event

REQ_APPLY_OPCODE

The wire opcode for this request

REQ_APPLY_SINCE

The minimal object version supporting this request

REQ_COLORCURVES_OPCODE

The wire opcode for this request

REQ_COLORCURVES_SINCE

The minimal object version supporting this request

REQ_DESTROY_OPCODE

The wire opcode for this request

REQ_DESTROY_SINCE

The minimal object version supporting this request

REQ_ENABLE_OPCODE

The wire opcode for this request

REQ_ENABLE_SINCE

The minimal object version supporting this request

REQ_MODE_OPCODE

The wire opcode for this request

REQ_MODE_SINCE

The minimal object version supporting this request

REQ_OVERSCAN_OPCODE

The wire opcode for this request

REQ_OVERSCAN_SINCE

The minimal object version supporting this request

REQ_POSITION_OPCODE

The wire opcode for this request

REQ_POSITION_SINCE

The minimal object version supporting this request

REQ_SCALEF_OPCODE

The wire opcode for this request

REQ_SCALEF_SINCE

The minimal object version supporting this request

REQ_SCALE_OPCODE

The wire opcode for this request

REQ_SCALE_SINCE

The minimal object version supporting this request

REQ_SET_VRR_POLICY_OPCODE

The wire opcode for this request

REQ_SET_VRR_POLICY_SINCE

The minimal object version supporting this request

REQ_TRANSFORM_OPCODE

The wire opcode for this request

REQ_TRANSFORM_SINCE

The minimal object version supporting this request