wayland_protocols_plasma::output_management::v2::client

Module kde_output_configuration_v2

configure single output devices

outputconfiguration is a client-specific resource that can be used to ask the server to apply changes to available output devices.

The client receives a list of output devices from the registry. When it wants to apply new settings, it creates a configuration object from the outputmanagement global, writes changes through this object’s enable, scale, transform and mode calls. It then asks the server to apply these settings in an atomic fashion, for example through Linux’ DRM interface.

The server signals back whether the new settings have applied successfully or failed to apply. outputdevice objects are updated after the changes have been applied to the hardware and before the server side sends the applied event.

Structs

KdeOutputConfigurationV2

configure single output devices

Enums

AutoRotatePolicy

describes when auto rotate should be used

ColorPowerTradeoff

tradeoff between power and accuracy

ColorProfileSource

which source the compositor should use for the color profile on an output

Error

kde_output_configuration_v2 error values

Event

Request

RgbRange

describes RGB range policy

VrrPolicy

describes vrr policy

Constants

EVT_APPLIED_OPCODE

The wire opcode for this event

EVT_APPLIED_SINCE

The minimal object version supporting this event

EVT_FAILED_OPCODE

The wire opcode for this event

EVT_FAILED_SINCE

The minimal object version supporting this event

EVT_FAILURE_REASON_OPCODE

The wire opcode for this event

EVT_FAILURE_REASON_SINCE

The minimal object version supporting this event

REQ_APPLY_OPCODE

The wire opcode for this request

REQ_APPLY_SINCE

The minimal object version supporting this request

REQ_DESTROY_OPCODE

The wire opcode for this request

REQ_DESTROY_SINCE

The minimal object version supporting this request

REQ_ENABLE_OPCODE

The wire opcode for this request

REQ_ENABLE_SINCE

The minimal object version supporting this request

REQ_MODE_OPCODE

The wire opcode for this request

REQ_MODE_SINCE

The minimal object version supporting this request

REQ_OVERSCAN_OPCODE

The wire opcode for this request

REQ_OVERSCAN_SINCE

The minimal object version supporting this request

REQ_POSITION_OPCODE

The wire opcode for this request

REQ_POSITION_SINCE

The minimal object version supporting this request

REQ_SCALE_OPCODE

The wire opcode for this request

REQ_SCALE_SINCE

The minimal object version supporting this request

REQ_SET_AUTO_ROTATE_POLICY_OPCODE

The wire opcode for this request

REQ_SET_AUTO_ROTATE_POLICY_SINCE

The minimal object version supporting this request

REQ_SET_BRIGHTNESS_OPCODE

The wire opcode for this request

REQ_SET_BRIGHTNESS_OVERRIDES_OPCODE

The wire opcode for this request

REQ_SET_BRIGHTNESS_OVERRIDES_SINCE

The minimal object version supporting this request

REQ_SET_BRIGHTNESS_SINCE

The minimal object version supporting this request

REQ_SET_COLOR_POWER_TRADEOFF_OPCODE

The wire opcode for this request

REQ_SET_COLOR_POWER_TRADEOFF_SINCE

The minimal object version supporting this request

REQ_SET_COLOR_PROFILE_SOURCE_OPCODE

The wire opcode for this request

REQ_SET_COLOR_PROFILE_SOURCE_SINCE

The minimal object version supporting this request

REQ_SET_DIMMING_OPCODE

The wire opcode for this request

REQ_SET_DIMMING_SINCE

The minimal object version supporting this request

REQ_SET_HIGH_DYNAMIC_RANGE_OPCODE

The wire opcode for this request

REQ_SET_HIGH_DYNAMIC_RANGE_SINCE

The minimal object version supporting this request

REQ_SET_ICC_PROFILE_PATH_OPCODE

The wire opcode for this request

REQ_SET_ICC_PROFILE_PATH_SINCE

The minimal object version supporting this request

REQ_SET_PRIMARY_OUTPUT_OPCODE

The wire opcode for this request

REQ_SET_PRIMARY_OUTPUT_SINCE

The minimal object version supporting this request

REQ_SET_PRIORITY_OPCODE

The wire opcode for this request

REQ_SET_PRIORITY_SINCE

The minimal object version supporting this request

REQ_SET_RGB_RANGE_OPCODE

The wire opcode for this request

REQ_SET_RGB_RANGE_SINCE

The minimal object version supporting this request

REQ_SET_SDR_BRIGHTNESS_OPCODE

The wire opcode for this request

REQ_SET_SDR_BRIGHTNESS_SINCE

The minimal object version supporting this request

REQ_SET_SDR_GAMUT_WIDENESS_OPCODE

The wire opcode for this request

REQ_SET_SDR_GAMUT_WIDENESS_SINCE

The minimal object version supporting this request

REQ_SET_VRR_POLICY_OPCODE

The wire opcode for this request

REQ_SET_VRR_POLICY_SINCE

The minimal object version supporting this request

REQ_SET_WIDE_COLOR_GAMUT_OPCODE

The wire opcode for this request

REQ_SET_WIDE_COLOR_GAMUT_SINCE

The minimal object version supporting this request

REQ_TRANSFORM_OPCODE

The wire opcode for this request

REQ_TRANSFORM_SINCE

The minimal object version supporting this request