wayland_protocols_plasma::output_order::v1::client

Module kde_output_order_v1

announce order of outputs

Announce the order in which desktop environment components should be placed on outputs. The compositor will send the list of outputs when the global is bound and whenever there is a change.

Warning! The protocol described in this file is a desktop environment implementation detail. Regular clients must not use this protocol. Backward incompatible changes may be added without bumping the major version of the extension.

Structs

KdeOutputOrderV1

announce order of outputs

Enums

Event

Request

Constants

EVT_DONE_OPCODE

The wire opcode for this event

EVT_DONE_SINCE

The minimal object version supporting this event

EVT_OUTPUT_OPCODE

The wire opcode for this event

EVT_OUTPUT_SINCE

The minimal object version supporting this event

REQ_DESTROY_OPCODE

The wire opcode for this request

REQ_DESTROY_SINCE

The minimal object version supporting this request