wayland_protocols_plasma::plasma_window_management::client::org_kde_plasma_window
Enum Request

#[non_exhaustive]
pub enum Request<'a> {

    SetState {
        flags: u32,
        state: u32,
    },
    SetVirtualDesktop {
        number: u32,
    },
    SetMinimizedGeometry {
        panel: WlSurface,
        x: u32,
        y: u32,
        width: u32,
        height: u32,
    },
    UnsetMinimizedGeometry {
        panel: WlSurface,
    },
    Close,
    RequestMove,
    RequestResize,
    Destroy,
    GetIcon {
        fd: BorrowedFd<'a>,
    },
    RequestEnterVirtualDesktop {
        id: String,
    },
    RequestEnterNewVirtualDesktop,
    RequestLeaveVirtualDesktop {
        id: String,
    },
    RequestEnterActivity {
        id: String,
    },
    RequestLeaveActivity {
        id: String,
    },
    SendToOutput {
        output: WlOutput,
    },

}

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

SetState

set window state

Set window state.

Values for state argument are described by org_kde_plasma_window_management.state and can be used together in a bitfield. The flags bitfield describes which flags are supposed to be set, the state bitfield the value for the set flags

Fields

flags: u32

bitfield of set state flags

state: u32

bitfield of state flags

SetVirtualDesktop

map window on a virtual desktop

Deprecated: use enter_virtual_desktop Maps the window to a different virtual desktop.

To show the window on all virtual desktops, call the org_kde_plasma_window.set_state request and specify a on_all_desktops state in the bitfield.

Fields

number: u32

zero based virtual desktop number

SetMinimizedGeometry

set the geometry for a taskbar entry

Sets the geometry of the taskbar entry for this window. The geometry is relative to a panel in particular.

Fields

panel: WlSurface

x: u32

y: u32

width: u32

height: u32

UnsetMinimizedGeometry

set the geometry for a taskbar entry

Remove the task geometry information for a particular panel.

Fields

panel: WlSurface

Close

close window

Close this window.

RequestMove

request move

Request an interactive move for this window.

Only available since version 3 of the interface

RequestResize

request resize

Request an interactive resize for this window.

Only available since version 3 of the interface

Destroy

remove resource for the org_kde_plasma_window

Removes the resource bound for this org_kde_plasma_window.

This is a destructor, once sent this object cannot be used any longer. Only available since version 4 of the interface

GetIcon

Requests to get the window icon

The compositor will write the window icon into the provided file descriptor. The data is a serialized QIcon with QDataStream.

Only available since version 7 of the interface

Fields

fd: BorrowedFd<'a>

file descriptor for the icon

RequestEnterVirtualDesktop

map window on a virtual desktop

Make the window enter a virtual desktop. A window can enter more than one virtual desktop. if the id is empty or invalid, no action will be performed.

Only available since version 8 of the interface

Fields

id: String

desktop id

RequestEnterNewVirtualDesktop

map window on a virtual desktop

RFC: do this with an empty id to request_enter_virtual_desktop? Make the window enter a new virtual desktop. If the server consents the request, it will create a new virtual desktop and assign the window to it.

Only available since version 8 of the interface

RequestLeaveVirtualDesktop

remove a window from a virtual desktop

Make the window exit a virtual desktop. If it exits all desktops it will be considered on all of them.

Only available since version 8 of the interface

Fields

id: String

desktop id

RequestEnterActivity

map window on an activity

Make the window enter an activity. A window can enter more activity. If the id is empty or invalid, no action will be performed.

Only available since version 14 of the interface

Fields

id: String

activity id

RequestLeaveActivity

remove a window from an activity

Make the window exit a an activity. If it exits all activities it will be considered on all of them.

Only available since version 14 of the interface

Fields

id: String

activity id

SendToOutput

send window to specified output

Requests this window to be displayed in a specific output.

Only available since version 15 of the interface

Fields

output: WlOutput

Implementations

impl<'a> Request<'a>

pub fn opcode(&self) -> u16

Get the opcode number of this message