wayland_protocols_plasma::screen_edge::v1::client

Module kde_screen_edge_manager_v1

screen edge manager

This interface allows clients to associate actions with screen edges. For example, showing a surface by moving the pointer to a screen edge.

Potential ways to trigger the screen edge are subject to compositor policies. As an example, the compositor may consider the screen edge to be triggered if the pointer hits its associated screen border. Other ways may include using touchscreen or touchpad gestures.

Warning! The protocol described in this file is a desktop environment implementation detail. Regular clients must not use this protocol. Backward incompatible changes may be added without bumping the major version of the extension.

Structs

KdeScreenEdgeManagerV1

screen edge manager

Enums

Border

screen border

Error

Event

Request

Constants

REQ_DESTROY_OPCODE

The wire opcode for this request

REQ_DESTROY_SINCE

The minimal object version supporting this request

REQ_GET_AUTO_HIDE_SCREEN_EDGE_OPCODE

The wire opcode for this request

REQ_GET_AUTO_HIDE_SCREEN_EDGE_SINCE

The minimal object version supporting this request