wayland_protocols_plasma::screencast::v1::client

Module zkde_screencast_unstable_v1

Protocol for managing PipeWire feeds of the different displays and windows

Warning! The protocol described in this file is a desktop environment implementation detail. Regular clients must not use this protocol. Backward incompatible changes may be added without bumping the major version of the extension.

Structs

ZkdeScreencastUnstableV1

Protocol for managing PipeWire feeds of the different displays and windows

Enums

Event

Pointer

Stream consumer attachment attributes

Request

Constants

REQ_DESTROY_OPCODE

The wire opcode for this request

REQ_DESTROY_SINCE

The minimal object version supporting this request

REQ_STREAM_OUTPUT_OPCODE

The wire opcode for this request

REQ_STREAM_OUTPUT_SINCE

The minimal object version supporting this request

REQ_STREAM_REGION_OPCODE

The wire opcode for this request

REQ_STREAM_REGION_SINCE

The minimal object version supporting this request

REQ_STREAM_VIRTUAL_OUTPUT_OPCODE

The wire opcode for this request

REQ_STREAM_VIRTUAL_OUTPUT_SINCE

The minimal object version supporting this request

REQ_STREAM_VIRTUAL_OUTPUT_WITH_DESCRIPTION_OPCODE

The wire opcode for this request

REQ_STREAM_VIRTUAL_OUTPUT_WITH_DESCRIPTION_SINCE

The minimal object version supporting this request

REQ_STREAM_WINDOW_OPCODE

The wire opcode for this request

REQ_STREAM_WINDOW_SINCE

The minimal object version supporting this request