wayland_protocols_plasma::server_decoration::client
Module org_kde_kwin_server_decoration_manager

Server side window decoration manager

This interface allows to coordinate whether the server should create a server-side window decoration around a wl_surface representing a shell surface (wl_shell_surface or similar). By announcing support for this interface the server indicates that it supports server side decorations.

Use in conjunction with zxdg_decoration_manager_v1 is undefined.

Structs

OrgKdeKwinServerDecorationManager

Server side window decoration manager

Enums

Event

Mode

Possible values to use in request_mode and the event mode.

Request

Constants

EVT_DEFAULT_MODE_OPCODE

The wire opcode for this event

EVT_DEFAULT_MODE_SINCE

The minimal object version supporting this event

REQ_CREATE_OPCODE

The wire opcode for this request

REQ_CREATE_SINCE

The minimal object version supporting this request