wayland_protocols_plasma::slide::client
Module org_kde_kwin_slide

slide a surface from a location to another

Ask the compositor to move the surface from a location to another with a slide animation.

The from argument provides a clue about where the slide animation begins, offset is the distance from screen edge to begin the animation.

Structs

OrgKdeKwinSlide

slide a surface from a location to another

Enums

Event

Location

Request

Constants

REQ_COMMIT_OPCODE

The wire opcode for this request

REQ_COMMIT_SINCE

The minimal object version supporting this request

REQ_RELEASE_OPCODE

The wire opcode for this request

REQ_RELEASE_SINCE

The minimal object version supporting this request

REQ_SET_LOCATION_OPCODE

The wire opcode for this request

REQ_SET_LOCATION_SINCE

The minimal object version supporting this request

REQ_SET_OFFSET_OPCODE

The wire opcode for this request

REQ_SET_OFFSET_SINCE

The minimal object version supporting this request