wayland_protocols_plasma::text_input::v2::client::zwp_text_input_v2
Enum Request

#[non_exhaustive]
pub enum Request<'a> {
    Destroy,
    Enable {
        surface: WlSurface,
    },
    Disable {
        surface: WlSurface,
    },
    ShowInputPanel,
    HideInputPanel,
    SetSurroundingText {
        text: String,
        cursor: i32,
        anchor: i32,
    },
    SetContentType {
        hint: WEnum<ContentHint>,
        purpose: WEnum<ContentPurpose>,
    },
    SetCursorRectangle {
        x: i32,
        y: i32,
        width: i32,
        height: i32,
    },
    SetPreferredLanguage {
        language: String,
    },
    UpdateState {
        serial: u32,
        reason: WEnum<UpdateState>,
    },
}

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Destroy

Destroy the wp_text_input

Destroy the wp_text_input object. Also disables all surfaces enabled through this wp_text_input object

This is a destructor, once sent this object cannot be used any longer.

Enable

enable text input for surface

Enable text input in a surface (usually when a text entry inside of it has focus).

This can be called before or after a surface gets text (or keyboard) focus via the enter event. Text input to a surface is only active when it has the current text (or keyboard) focus and is enabled.

Fields

surface: WlSurface

Disable

disable text input for surface

Disable text input in a surface (typically when there is no focus on any text entry inside the surface).

Fields

surface: WlSurface

ShowInputPanel

show input panels

Requests input panels (virtual keyboard) to show.

This should be used for example to show a virtual keyboard again (with a tap) after it was closed by pressing on a close button on the keyboard.

HideInputPanel

hide input panels

Requests input panels (virtual keyboard) to hide.

SetSurroundingText

sets the surrounding text

Sets the plain surrounding text around the input position. Text is UTF-8 encoded. Cursor is the byte offset within the surrounding text. Anchor is the byte offset of the selection anchor within the surrounding text. If there is no selected text, anchor is the same as cursor.

Make sure to always send some text before and after the cursor except when the cursor is at the beginning or end of text.

When there was a configure_surrounding_text event take the before_cursor and after_cursor arguments into account for picking how much surrounding text to send.

There is a maximum length of wayland messages so text can not be longer than 4000 bytes.

Fields

text: String

cursor: i32

anchor: i32

SetContentType

set content purpose and hint

Sets the content purpose and content hint. While the purpose is the basic purpose of an input field, the hint flags allow to modify some of the behavior.

When no content type is explicitly set, a normal content purpose with none hint should be assumed.

Fields

hint: WEnum<ContentHint>

purpose: WEnum<ContentPurpose>

SetCursorRectangle

set cursor position

Sets the cursor outline as a x, y, width, height rectangle in surface local coordinates.

Allows the compositor to put a window with word suggestions near the cursor.

Fields

x: i32

y: i32

width: i32

height: i32

SetPreferredLanguage

sets preferred language

Sets a specific language. This allows for example a virtual keyboard to show a language specific layout. The “language” argument is a RFC-3066 format language tag.

It could be used for example in a word processor to indicate language of currently edited document or in an instant message application which tracks languages of contacts.

Fields

language: String

UpdateState

update state

Allows to atomically send state updates from client.

This request should follow after a batch of state updating requests like set_surrounding_text, set_content_type, set_cursor_rectangle and set_preferred_language.

The flags field indicates why an updated state is sent to the input method.

Reset should be used by an editor widget after the text was changed outside of the normal input method flow.

For “change” it is enough to send the changed state, else the full state should be send.

Serial should be set to the serial from the last enter or input_method_changed event.

To make sure to not receive outdated input method events after a reset or switching to a new widget wl_display_sync() should be used after update_state in these cases.

Fields

serial: u32

serial of the enter or input_method_changed event

reason: WEnum<UpdateState>

Implementations

impl<'a> Request<'a>

pub fn opcode(&self) -> u16

Get the opcode number of this message