Module wayland_protocols_plasma::text_input::v2::server::zwp_text_input_v2

text input

The zwp_text_input_v2 interface represents text input and input methods associated with a seat. It provides enter/leave events to follow the text input focus for a seat.

Requests are used to enable/disable the text-input object and set state information like surrounding and selected text or the content type. The information about the entered text is sent to the text-input object via the pre-edit and commit events. Using this interface removes the need for applications to directly process hardware key events and compose text out of them.

Text is valid UTF-8 encoded, indices and lengths are in bytes. Indices have to always point to the first byte of an UTF-8 encoded code point. Lengths are not allowed to contain just a part of an UTF-8 encoded code point.

State is sent by the state requests (set_surrounding_text, set_content_type, set_cursor_rectangle and set_preferred_language) and an update_state request. After an enter or an input_method_change event all state information is invalidated and needs to be resent from the client. A reset or entering a new widget on client side also invalidates all current state information.

Structs

ContentHint

content hint

ZwpTextInputV2

text input

Enums

ContentPurpose

content purpose

Event

InputPanelVisibility

PreeditStyle

Request

TextDirection

UpdateState

update_state flags

Constants

EVT_COMMIT_STRING_OPCODE

The wire opcode for this event

EVT_COMMIT_STRING_SINCE

The minimal object version supporting this event

EVT_CONFIGURE_SURROUNDING_TEXT_OPCODE

The wire opcode for this event

EVT_CONFIGURE_SURROUNDING_TEXT_SINCE

The minimal object version supporting this event

EVT_CURSOR_POSITION_OPCODE

The wire opcode for this event

EVT_CURSOR_POSITION_SINCE

The minimal object version supporting this event

EVT_DELETE_SURROUNDING_TEXT_OPCODE

The wire opcode for this event

EVT_DELETE_SURROUNDING_TEXT_SINCE

The minimal object version supporting this event

EVT_ENTER_OPCODE

The wire opcode for this event

EVT_ENTER_SINCE

The minimal object version supporting this event

EVT_INPUT_METHOD_CHANGED_OPCODE

The wire opcode for this event

EVT_INPUT_METHOD_CHANGED_SINCE

The minimal object version supporting this event

EVT_INPUT_PANEL_STATE_OPCODE

The wire opcode for this event

EVT_INPUT_PANEL_STATE_SINCE

The minimal object version supporting this event

EVT_KEYSYM_OPCODE

The wire opcode for this event

EVT_KEYSYM_SINCE

The minimal object version supporting this event

EVT_LANGUAGE_OPCODE

The wire opcode for this event

EVT_LANGUAGE_SINCE

The minimal object version supporting this event

EVT_LEAVE_OPCODE

The wire opcode for this event

EVT_LEAVE_SINCE

The minimal object version supporting this event

EVT_MODIFIERS_MAP_OPCODE

The wire opcode for this event

EVT_MODIFIERS_MAP_SINCE

The minimal object version supporting this event

EVT_PREEDIT_CURSOR_OPCODE

The wire opcode for this event

EVT_PREEDIT_CURSOR_SINCE

The minimal object version supporting this event

EVT_PREEDIT_STRING_OPCODE

The wire opcode for this event

EVT_PREEDIT_STRING_SINCE

The minimal object version supporting this event

EVT_PREEDIT_STYLING_OPCODE

The wire opcode for this event

EVT_PREEDIT_STYLING_SINCE

The minimal object version supporting this event

EVT_TEXT_DIRECTION_OPCODE

The wire opcode for this event

EVT_TEXT_DIRECTION_SINCE

The minimal object version supporting this event

REQ_DESTROY_OPCODE

The wire opcode for this request

REQ_DESTROY_SINCE

The minimal object version supporting this request

REQ_DISABLE_OPCODE

The wire opcode for this request

REQ_DISABLE_SINCE

The minimal object version supporting this request

REQ_ENABLE_OPCODE

The wire opcode for this request

REQ_ENABLE_SINCE

The minimal object version supporting this request

REQ_HIDE_INPUT_PANEL_OPCODE

The wire opcode for this request

REQ_HIDE_INPUT_PANEL_SINCE

The minimal object version supporting this request

REQ_SET_CONTENT_TYPE_OPCODE

The wire opcode for this request

REQ_SET_CONTENT_TYPE_SINCE

The minimal object version supporting this request

REQ_SET_CURSOR_RECTANGLE_OPCODE

The wire opcode for this request

REQ_SET_CURSOR_RECTANGLE_SINCE

The minimal object version supporting this request

REQ_SET_PREFERRED_LANGUAGE_OPCODE

The wire opcode for this request

REQ_SET_PREFERRED_LANGUAGE_SINCE

The minimal object version supporting this request

REQ_SET_SURROUNDING_TEXT_OPCODE

The wire opcode for this request

REQ_SET_SURROUNDING_TEXT_SINCE

The minimal object version supporting this request

REQ_SHOW_INPUT_PANEL_OPCODE

The wire opcode for this request

REQ_SHOW_INPUT_PANEL_SINCE

The minimal object version supporting this request

REQ_UPDATE_STATE_OPCODE

The wire opcode for this request

REQ_UPDATE_STATE_SINCE

The minimal object version supporting this request