Crate wayland_protocols_plasma

This crate provides bindings to the plasma wayland protocol extensions provided in https://github.com/KDE/plasma-wayland-protocols

These bindings are built on top of the crates wayland-client and wayland-server.

Each protocol module contains a client and a server submodules, for each side of the protocol. The creation of these modules (and the dependency on the associated crate) is controlled by the two cargo features client and server.

Modules

appmenu

blur

contrast

dpms

fake_input

idle

keystate

output_device

output_management

plasma_shell

plasma_virtual_desktop

primary_output

remote_access

screencast

server_decoration

server_decoration_palette

shadow

slide

text_input

wayland_eglstream_controller