wayland_protocols_plasma::output_device::v2::client

Module kde_output_device_v2

output configuration representation

An output device describes a display device available to the compositor. output_device is similar to wl_output, but focuses on output configuration management.

A client can query all global output_device objects to enlist all available display devices, even those that may currently not be represented by the compositor as a wl_output.

The client sends configuration changes to the server through the outputconfiguration interface, and the server applies the configuration changes to the hardware and signals changes to the output devices accordingly.

This object is published as global during start up for every available display devices, or when one later becomes available, for example by being hotplugged via a physical connector.

Warning! The protocol described in this file is a desktop environment implementation detail. Regular clients must not use this protocol. Backward incompatible changes may be added without bumping the major version of the extension.

Structs

Capability

describes capabilities of the outputdevice

KdeOutputDeviceV2

output configuration representation

Enums

AutoRotatePolicy

describes when auto rotate should be used

ColorPowerTradeoff

tradeoff between power and accuracy

ColorProfileSource

which source the compositor should use for the color profile on an output

Event

Request

RgbRange

describes RGB range policy

Subpixel

subpixel geometry information

Transform

transform from framebuffer to output

VrrPolicy

describes vrr policy

Constants

EVT_AUTO_ROTATE_POLICY_OPCODE

The wire opcode for this event

EVT_AUTO_ROTATE_POLICY_SINCE

The minimal object version supporting this event

EVT_BRIGHTNESS_METADATA_OPCODE

The wire opcode for this event

EVT_BRIGHTNESS_METADATA_SINCE

The minimal object version supporting this event

EVT_BRIGHTNESS_OPCODE

The wire opcode for this event

EVT_BRIGHTNESS_OVERRIDES_OPCODE

The wire opcode for this event

EVT_BRIGHTNESS_OVERRIDES_SINCE

The minimal object version supporting this event

EVT_BRIGHTNESS_SINCE

The minimal object version supporting this event

EVT_CAPABILITIES_OPCODE

The wire opcode for this event

EVT_CAPABILITIES_SINCE

The minimal object version supporting this event

EVT_COLOR_POWER_TRADEOFF_OPCODE

The wire opcode for this event

EVT_COLOR_POWER_TRADEOFF_SINCE

The minimal object version supporting this event

EVT_COLOR_PROFILE_SOURCE_OPCODE

The wire opcode for this event

EVT_COLOR_PROFILE_SOURCE_SINCE

The minimal object version supporting this event

EVT_CURRENT_MODE_OPCODE

The wire opcode for this event

EVT_CURRENT_MODE_SINCE

The minimal object version supporting this event

EVT_DIMMING_OPCODE

The wire opcode for this event

EVT_DIMMING_SINCE

The minimal object version supporting this event

EVT_DONE_OPCODE

The wire opcode for this event

EVT_DONE_SINCE

The minimal object version supporting this event

EVT_EDID_OPCODE

The wire opcode for this event

EVT_EDID_SINCE

The minimal object version supporting this event

EVT_EISA_ID_OPCODE

The wire opcode for this event

EVT_EISA_ID_SINCE

The minimal object version supporting this event

EVT_ENABLED_OPCODE

The wire opcode for this event

EVT_ENABLED_SINCE

The minimal object version supporting this event

EVT_GEOMETRY_OPCODE

The wire opcode for this event

EVT_GEOMETRY_SINCE

The minimal object version supporting this event

EVT_HIGH_DYNAMIC_RANGE_OPCODE

The wire opcode for this event

EVT_HIGH_DYNAMIC_RANGE_SINCE

The minimal object version supporting this event

EVT_ICC_PROFILE_PATH_OPCODE

The wire opcode for this event

EVT_ICC_PROFILE_PATH_SINCE

The minimal object version supporting this event

EVT_MODE_OPCODE

The wire opcode for this event

EVT_MODE_SINCE

The minimal object version supporting this event

EVT_NAME_OPCODE

The wire opcode for this event

EVT_NAME_SINCE

The minimal object version supporting this event

EVT_OVERSCAN_OPCODE

The wire opcode for this event

EVT_OVERSCAN_SINCE

The minimal object version supporting this event

EVT_RGB_RANGE_OPCODE

The wire opcode for this event

EVT_RGB_RANGE_SINCE

The minimal object version supporting this event

EVT_SCALE_OPCODE

The wire opcode for this event

EVT_SCALE_SINCE

The minimal object version supporting this event

EVT_SDR_BRIGHTNESS_OPCODE

The wire opcode for this event

EVT_SDR_BRIGHTNESS_SINCE

The minimal object version supporting this event

EVT_SDR_GAMUT_WIDENESS_OPCODE

The wire opcode for this event

EVT_SDR_GAMUT_WIDENESS_SINCE

The minimal object version supporting this event

EVT_SERIAL_NUMBER_OPCODE

The wire opcode for this event

EVT_SERIAL_NUMBER_SINCE

The minimal object version supporting this event

EVT_UUID_OPCODE

The wire opcode for this event

EVT_UUID_SINCE

The minimal object version supporting this event

EVT_VRR_POLICY_OPCODE

The wire opcode for this event

EVT_VRR_POLICY_SINCE

The minimal object version supporting this event

EVT_WIDE_COLOR_GAMUT_OPCODE

The wire opcode for this event

EVT_WIDE_COLOR_GAMUT_SINCE

The minimal object version supporting this event