Module wayland_protocols_plasma::output_management::v2::client::kde_output_management_v2

configuration of server outputs through clients

This interface enables clients to set properties of output devices for screen configuration purposes via the server. To this end output devices are referenced by global kde_output_device_v2 objects.

outputmanagement (wl_global)

request:

create_configuration -> outputconfiguration (wl_resource)

outputconfiguration (wl_resource)

requests:

enable(outputdevice, bool)

mode(outputdevice, mode)

transformation(outputdevice, flag)

position(outputdevice, x, y)

apply

events:

applied

failed

The server registers one outputmanagement object as a global object. In order to configure outputs a client requests create_configuration, which provides a resource referencing an outputconfiguration for one-time configuration. That way the server knows which requests belong together and can group them by that.

On the outputconfiguration object the client calls for each output whether the output should be enabled, which mode should be set (by referencing the mode from the list of announced modes) and the output’s global position. Once all outputs are configured that way, the client calls apply. At that point and not earlier the server should try to apply the configuration. If this succeeds the server emits the applied signal, otherwise the failed signal, such that the configuring client is noticed about the success of its configuration request.

Through this design the interface enables atomic output configuration changes if internally supported by the server.

Structs

KdeOutputManagementV2

configuration of server outputs through clients

Enums

Event

Request

Constants

REQ_CREATE_CONFIGURATION_OPCODE

The wire opcode for this request

REQ_CREATE_CONFIGURATION_SINCE

The minimal object version supporting this request