wayland_protocols_plasma::plasma_shell::client

Module org_kde_plasma_surface

metadata interface

An interface that may be implemented by a wl_surface, for implementations that provide the shell user interface.

It provides requests to set surface roles, assign an output or set the position in output coordinates.

On the server side the object is automatically destroyed when the related wl_surface is destroyed. On client side, org_kde_plasma_surface.destroy() must be called before destroying the wl_surface object.

Structs

OrgKdePlasmaSurface

metadata interface

Enums

Error

Event

PanelBehavior

Behavior for panel surface

Request

Role

Constants

EVT_AUTO_HIDDEN_PANEL_HIDDEN_OPCODE

The wire opcode for this event

EVT_AUTO_HIDDEN_PANEL_HIDDEN_SINCE

The minimal object version supporting this event

EVT_AUTO_HIDDEN_PANEL_SHOWN_OPCODE

The wire opcode for this event

EVT_AUTO_HIDDEN_PANEL_SHOWN_SINCE

The minimal object version supporting this event

REQ_DESTROY_OPCODE

The wire opcode for this request

REQ_DESTROY_SINCE

The minimal object version supporting this request

REQ_OPEN_UNDER_CURSOR_OPCODE

The wire opcode for this request

REQ_OPEN_UNDER_CURSOR_SINCE

The minimal object version supporting this request

REQ_PANEL_AUTO_HIDE_HIDE_OPCODE

The wire opcode for this request

REQ_PANEL_AUTO_HIDE_HIDE_SINCE

The minimal object version supporting this request

REQ_PANEL_AUTO_HIDE_SHOW_OPCODE

The wire opcode for this request

REQ_PANEL_AUTO_HIDE_SHOW_SINCE

The minimal object version supporting this request

REQ_SET_OUTPUT_OPCODE

The wire opcode for this request

REQ_SET_OUTPUT_SINCE

The minimal object version supporting this request

REQ_SET_PANEL_BEHAVIOR_OPCODE

The wire opcode for this request

REQ_SET_PANEL_BEHAVIOR_SINCE

The minimal object version supporting this request

REQ_SET_PANEL_TAKES_FOCUS_OPCODE

The wire opcode for this request

REQ_SET_PANEL_TAKES_FOCUS_SINCE

The minimal object version supporting this request

REQ_SET_POSITION_OPCODE

The wire opcode for this request

REQ_SET_POSITION_SINCE

The minimal object version supporting this request

REQ_SET_ROLE_OPCODE

The wire opcode for this request

REQ_SET_ROLE_SINCE

The minimal object version supporting this request

REQ_SET_SKIP_SWITCHER_OPCODE

The wire opcode for this request

REQ_SET_SKIP_SWITCHER_SINCE

The minimal object version supporting this request

REQ_SET_SKIP_TASKBAR_OPCODE

The wire opcode for this request

REQ_SET_SKIP_TASKBAR_SINCE

The minimal object version supporting this request