wayland_protocols_plasma::plasma_window_management::client

Module org_kde_plasma_window_management

application windows management

This interface manages application windows. It provides requests to show and hide the desktop and emits an event every time a window is created so that the client can use it to manage the window.

Only one client can bind this interface at a time.

Warning! The protocol described in this file is a desktop environment implementation detail. Regular clients must not use this protocol. Backward incompatible changes may be added without bumping the major version of the extension.

Structs

OrgKdePlasmaWindowManagement

application windows management

Enums

Event

Request

ShowDesktop

State

Constants

EVT_SHOW_DESKTOP_CHANGED_OPCODE

The wire opcode for this event

EVT_SHOW_DESKTOP_CHANGED_SINCE

The minimal object version supporting this event

EVT_STACKING_ORDER_CHANGED_2_OPCODE

The wire opcode for this event

EVT_STACKING_ORDER_CHANGED_2_SINCE

The minimal object version supporting this event

EVT_STACKING_ORDER_CHANGED_OPCODE

The wire opcode for this event

EVT_STACKING_ORDER_CHANGED_SINCE

The minimal object version supporting this event

EVT_STACKING_ORDER_UUID_CHANGED_OPCODE

The wire opcode for this event

EVT_STACKING_ORDER_UUID_CHANGED_SINCE

The minimal object version supporting this event

EVT_WINDOW_OPCODE

The wire opcode for this event

EVT_WINDOW_SINCE

The minimal object version supporting this event

EVT_WINDOW_WITH_UUID_OPCODE

The wire opcode for this event

EVT_WINDOW_WITH_UUID_SINCE

The minimal object version supporting this event

REQ_GET_STACKING_ORDER_OPCODE

The wire opcode for this request

REQ_GET_STACKING_ORDER_SINCE

The minimal object version supporting this request

REQ_GET_WINDOW_BY_UUID_OPCODE

The wire opcode for this request

REQ_GET_WINDOW_BY_UUID_SINCE

The minimal object version supporting this request

REQ_GET_WINDOW_OPCODE

The wire opcode for this request

REQ_GET_WINDOW_SINCE

The minimal object version supporting this request

REQ_SHOW_DESKTOP_OPCODE

The wire opcode for this request

REQ_SHOW_DESKTOP_SINCE

The minimal object version supporting this request