Module wayland_protocols_plasma::text_input::v1::server::wl_text_input

text input

An object used for text input. Adds support for text input and input methods to applications. A text-input object is created from a wl_text_input_manager and corresponds typically to a text entry in an application. Requests are used to activate/deactivate the text-input object and set state information like surrounding and selected text or the content type. The information about entered text is sent to the text-input object via the pre-edit and commit events. Using this interface removes the need for applications to directly process hardware key events and compose text out of them.

Text is generally UTF-8 encoded, indices and lengths are in bytes.

Serials are used to synchronize the state between the text input and an input method. New serials are sent by the text input in the commit_state request and are used by the input method to indicate the known text input state in events like preedit_string, commit_string, and keysym. The text input can then ignore events from the input method which are based on an outdated state (for example after a reset).

Structs

WlTextInput

text input

Enums

ContentHint

content hint

ContentPurpose

content purpose

Event

PreeditStyle

Request

TextDirection

Constants

EVT_COMMIT_STRING_OPCODE

The wire opcode for this event

EVT_COMMIT_STRING_SINCE

The minimal object version supporting this event

EVT_CURSOR_POSITION_OPCODE

The wire opcode for this event

EVT_CURSOR_POSITION_SINCE

The minimal object version supporting this event

EVT_DELETE_SURROUNDING_TEXT_OPCODE

The wire opcode for this event

EVT_DELETE_SURROUNDING_TEXT_SINCE

The minimal object version supporting this event

EVT_ENTER_OPCODE

The wire opcode for this event

EVT_ENTER_SINCE

The minimal object version supporting this event

EVT_INPUT_PANEL_STATE_OPCODE

The wire opcode for this event

EVT_INPUT_PANEL_STATE_SINCE

The minimal object version supporting this event

EVT_KEYSYM_OPCODE

The wire opcode for this event

EVT_KEYSYM_SINCE

The minimal object version supporting this event

EVT_LANGUAGE_OPCODE

The wire opcode for this event

EVT_LANGUAGE_SINCE

The minimal object version supporting this event

EVT_LEAVE_OPCODE

The wire opcode for this event

EVT_LEAVE_SINCE

The minimal object version supporting this event

EVT_MODIFIERS_MAP_OPCODE

The wire opcode for this event

EVT_MODIFIERS_MAP_SINCE

The minimal object version supporting this event

EVT_PREEDIT_CURSOR_OPCODE

The wire opcode for this event

EVT_PREEDIT_CURSOR_SINCE

The minimal object version supporting this event

EVT_PREEDIT_STRING_OPCODE

The wire opcode for this event

EVT_PREEDIT_STRING_SINCE

The minimal object version supporting this event

EVT_PREEDIT_STYLING_OPCODE

The wire opcode for this event

EVT_PREEDIT_STYLING_SINCE

The minimal object version supporting this event

EVT_TEXT_DIRECTION_OPCODE

The wire opcode for this event

EVT_TEXT_DIRECTION_SINCE

The minimal object version supporting this event

REQ_ACTIVATE_OPCODE

The wire opcode for this request

REQ_ACTIVATE_SINCE

The minimal object version supporting this request

REQ_COMMIT_STATE_OPCODE

The wire opcode for this request

REQ_COMMIT_STATE_SINCE

The minimal object version supporting this request

REQ_DEACTIVATE_OPCODE

The wire opcode for this request

REQ_DEACTIVATE_SINCE

The minimal object version supporting this request

REQ_HIDE_INPUT_PANEL_OPCODE

The wire opcode for this request

REQ_HIDE_INPUT_PANEL_SINCE

The minimal object version supporting this request

REQ_INVOKE_ACTION_OPCODE

The wire opcode for this request

REQ_INVOKE_ACTION_SINCE

The minimal object version supporting this request

REQ_RESET_OPCODE

The wire opcode for this request

REQ_RESET_SINCE

The minimal object version supporting this request

REQ_SET_CONTENT_TYPE_OPCODE

The wire opcode for this request

REQ_SET_CONTENT_TYPE_SINCE

The minimal object version supporting this request

REQ_SET_CURSOR_RECTANGLE_OPCODE

The wire opcode for this request

REQ_SET_CURSOR_RECTANGLE_SINCE

The minimal object version supporting this request

REQ_SET_PREFERRED_LANGUAGE_OPCODE

The wire opcode for this request

REQ_SET_PREFERRED_LANGUAGE_SINCE

The minimal object version supporting this request

REQ_SET_SURROUNDING_TEXT_OPCODE

The wire opcode for this request

REQ_SET_SURROUNDING_TEXT_SINCE

The minimal object version supporting this request

REQ_SHOW_INPUT_PANEL_OPCODE

The wire opcode for this request

REQ_SHOW_INPUT_PANEL_SINCE

The minimal object version supporting this request